+++ CMakeLists.txt
@@ -126,7 +126,6 @@ file(GLOB_RECURSE SRC_FILES
         src/core/shaders/*.*
         src/core/shapes/*.*
         src/core/utils/*.*
-        src/core/custom/*.*
         src/gpu/ops/*.*
         src/gpu/processors/*.*
         src/gpu/proxies/*.*
