+++ src/core/CustomTypeFaceBuilder.cpp
