--- src/core/custom/GlyphPathProvider.cpp
@@ -1,33 +1,0 @@
-/////////////////////////////////////////////////////////////////////////////////////////////////
-//
-//  Tencent is pleased to support the open source community by making tgfx available.
-//
-//  Copyright (C) 2025 THL A29 Limited, a Tencent company. All rights reserved.
-//
-//  Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
-//  in compliance with the License. You may obtain a copy of the License at
-//
-//      https://opensource.org/licenses/BSD-3-Clause
-//
-//  unless required by applicable law or agreed to in writing, software distributed under the
-//  license is distributed on an "as is" basis, without warranties or conditions of any kind,
-//  either express or implied. see the license for the specific language governing permissions
-//  and limitations under the license.
-//
-/////////////////////////////////////////////////////////////////////////////////////////////////
-
-#include "GlyphPathProvider.h"
-
-namespace tgfx {
-std::shared_ptr<PathProvider> GlyphPathProvider::Wrap(const Path& path) {
-  return std::shared_ptr<PathProvider>(new GlyphPathProvider(path));
-}
-
-Path GlyphPathProvider::getPath() const {
-  return _path;
-}
-
-Rect GlyphPathProvider::getBounds() const {
-  return _path.getBounds();
-}
-}  // namespace tgfx
