+++ src/core/vectors/freetype/FTTypeface.h
@@ -59,6 +59,8 @@ class FTTypeface : public Typeface {
   std::vector<Unichar> getGlyphToUnicodeMap() const override;
 #endif
 
+  std::shared_ptr<ScalerContext> onCreateScalerContext(float size) const override;
+
  private:
   uint32_t _uniqueID = 0;
   FTFontData data;
@@ -67,8 +69,6 @@ class FTTypeface : public Typeface {
   FTTypeface(FTFontData data, FT_Face face);
 
   int unitsPerEmInternal() const;
-
-  std::shared_ptr<ScalerContext> onCreateScalerContext(float size) const override;
 
   friend class FTScalerContext;
 };
